Split out HTTPRequest and RequestManager from HTTPClient. HTTPRequest->HTTPRequestBase.
