+++ Common/Common.vcxproj
@@ -453,12 +453,13 @@
     <ClInclude Include="Math\Statistics.h" />
     <ClInclude Include="Net\Connection.h" />
     <ClInclude Include="Net\HTTPNaettRequest.h" />
+    <ClInclude Include="net\HTTPRequest.h" />
     <ClInclude Include="Net\HTTPRequestManager.h" />
     <ClInclude Include="Net\NetBuffer.h" />
     <ClInclude Include="Net\HTTPClient.h" />
     <ClInclude Include="Net\HTTPHeaders.h" />
     <ClInclude Include="Net\HTTPServer.h" />
-    <ClInclude Include="Net\HTTPRequest.h" />
+    <ClInclude Include="Net\HTTPRequestBase.h" />
     <ClInclude Include="Net\Resolve.h" />
     <ClInclude Include="Net\Sinks.h" />
     <ClInclude Include="Net\SocketCompat.h" />
@@ -910,12 +911,13 @@
     <ClCompile Include="Math\Statistics.cpp" />
     <ClCompile Include="Net\Connection.cpp" />
     <ClCompile Include="Net\HTTPNaettRequest.cpp" />
+    <ClCompile Include="net\HTTPRequest.cpp" />
     <ClCompile Include="Net\HTTPRequestManager.cpp" />
     <ClCompile Include="Net\NetBuffer.cpp" />
     <ClCompile Include="Net\HTTPClient.cpp" />
     <ClCompile Include="Net\HTTPHeaders.cpp" />
     <ClCompile Include="Net\HTTPServer.cpp" />
-    <ClCompile Include="Net\HTTPRequest.cpp" />
+    <ClCompile Include="Net\HTTPRequestBase.cpp" />
     <ClCompile Include="Net\Resolve.cpp" />
     <ClCompile Include="Net\Sinks.cpp" />
     <ClCompile Include="Net\URL.cpp" />
