+++ Common/Net/HTTPClient.h
@@ -9,8 +9,6 @@
 #include "Common/File/Path.h"
 #include "Common/Net/NetBuffer.h"
 #include "Common/Net/Resolve.h"
-#include "Common/Net/HTTPRequest.h"
-
 #include "Common/Net/Connection.h"
 
 namespace http {
@@ -64,32 +62,6 @@ protected:
 	std::string userAgent_;
 	const char* httpVersion_;
 	double dataTimeout_ = 900.0;
-};
-
-// Really an asynchronous request.
-class HTTPRequest : public Request {
-public:
-	HTTPRequest(RequestMethod method, std::string_view url, std::string_view postData, std::string_view postMime, const Path &outfile, RequestFlags flags, net::ResolveFunc customResolve, std::string_view name = "");
-	~HTTPRequest();
-
-	void Start() override;
-	void Join() override;
-
-	bool Done() override { return completed_; }
-	bool Failed() const override { return failed_; }
-
-private:
-	void Do();  // Actually does the download. Runs on thread.
-	int Perform(const std::string &url);
-	std::string RedirectLocation(const std::string &baseUrl) const;
-	void SetFailed(int code);
-
-	std::string postData_;
-	std::thread thread_;
-	std::string postMime_;
-	bool completed_ = false;
-	bool failed_ = false;
-	net::ResolveFunc customResolve_;
 };
 
 }  // namespace http
